Rough outline of code
Super duper rough outline. The interrupts/event handlers work as is, just press ctrl-c once when the code is running to exit the first try/except conditional, then press buttons connected to the GPIO pins and the corresponding button function will print the function. i.e if the reset button is pressed, it'll print 're', same for the other buttons.